chore: select event trigger only once
